Change int to big int

+++ backend_auth/users/models.py
@@ -32,7 +32,7 @@ class UserManager(BaseUserManager):
 
 
 class User(AbstractBaseUser, PermissionsMixin):
-    user_id = models.IntegerField(primary_key=True, blank=False)
+    user_id = models.BigIntegerField(primary_key=True, blank=False)
     login = models.CharField(max_length=50, unique=True, verbose_name="Логин", null=True, blank=True)
     password = models.CharField(max_length=100, verbose_name="Пароль", null=True, blank=True)
     is_active = models.BooleanField(default=False, verbose_name="Активный")
